% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateSpan.R
\name{collapseDateSpan}
\alias{collapseDateSpan}
\title{Generates non overlapping date spans from date span.}
\usage{
collapseDateSpan(
  x,
  startDate = "startDate",
  endDate = "endDate",
  gap = 1,
  group = NULL
)
}
\arguments{
\item{x}{A dataframe with atleast two date columns called startDate, endDate.}

\item{startDate}{A date column in the dataframe x, that refers to the startDate. Default = 'startDate'}

\item{endDate}{A date column in the dataframe x, that refers to the endDate. Default = 'endDate'}

\item{gap}{(Optional) Default = 1.}

\item{group}{(optional) if there is a grouping varible.}
}
\value{
A tibble with two date columns startDate, endDate. if a group variable
                     is provided, it will be part of returned object.
}
\description{
This function takes an input data frame object with two date columns 
corresponding to startDate and endDate, where endDate >= startDate. 
It outputs a dataframe with non overlapping startDate and endDate - both dates
inclusive. Optionally, the input and output data frame may be grouped by a group
variable.
}
\examples{
\dontrun{
#If your data is in database, provide connection.
connection <- collapseDateSpan(x = dataFrame,
                               startDate = start,
                               endDate = end)
}
}
