% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommaSeparatedStringToIntArray.R
\name{commaSeparatedStringToIntArray}
\alias{commaSeparatedStringToIntArray}
\title{Convert a Comma-Separated String to a Numeric Vector}
\usage{
commaSeparatedStringToIntArray(inputString)
}
\arguments{
\item{inputString}{A character string containing numeric values separated by commas.}
}
\value{
A numeric vector with each element corresponding to a number extracted from the input string.
}
\description{
This function takes a comma-separated string, splits it into individual elements,
removes any empty components, and converts the remaining elements into numeric values.
}
\examples{
commaSeparatedStringToIntArray("1,2,3,4")
# [1] 1 2 3 4

}
