% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateSpan.R
\name{convertDateSpanToDateVector}
\alias{convertDateSpanToDateVector}
\title{Generates a vector of dates from date span}
\usage{
convertDateSpanToDateVector(x, startDate = "startDate", endDate = "endDate")
}
\arguments{
\item{x}{A dataframe with atleast two date columns called startDate, endDate.}

\item{startDate}{A date column in the dataframe x, that refers to the startDate. Default = 'startDate'}

\item{endDate}{A date column in the dataframe x, that refers to the endDate. Default = 'endDate'}
}
\value{
A vector of dates.
}
\description{
This function takes an input a data frame object with atleast two date columns 
corresponding to startDate and endDate, where endDate >= startDate. 
It outputs a vector of unique dates.
}
\examples{
\dontrun{
connection <- convertDateSpanToDateVector(x = dataFrame)
}
}
