% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateSpan.R
\name{convertDateVectorToDateSpan}
\alias{convertDateVectorToDateSpan}
\title{Converts a vector of dates into date spans}
\usage{
convertDateVectorToDateSpan(x, unit = NULL, week_start = NULL)
}
\arguments{
\item{x}{A vector of dates.}

\item{unit}{(Optional) A character string specifying a calendar unit such as 
day, week, month, quarter and year. Options are from 
\code{lubridate::floor_date}.}

\item{week_start}{(Optional) when unit is weeks, specify the reference day. 7 (default) 
represents Sunday and 1 represents Monday. Options from 
\code{lubridate::floor_date}.}
}
\value{
A tibble with two date columns startDate, endDate.
}
\description{
This function takes an input vector of dates and outputs a data frame (tibble) with
date spans represented by startDate and endDate - both dates inclusive. Optionally, 
the date spans maybe rounded using the parameters and conventions in \code{lubridate::floor_date}.
}
\examples{
\dontrun{
connection <- convertDateVectorToDateSpan(x = dataFrame,
                                          date = date)
}
}
