% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeFeatureExtraction.R
\name{executeFeatureExtraction}
\alias{executeFeatureExtraction}
\title{Execute feature extraction}
\usage{
executeFeatureExtraction(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortIds,
  cohortTable,
  covariateSettings = NULL,
  addCohortBasedTemporalCovariateSettings = TRUE,
  covariateCohortDatabaseSchema = cohortDatabaseSchema,
  includeCovariateIds = NULL,
  covariateCohortTable = cohortTable,
  covariateCohortDefinitionSet = NULL,
  cohortCovariateAnalysisId = 150,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outputFolder,
  aggregated = TRUE,
  rowIdField = "subject_id",
  incremental = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema with instantiated target cohorts.}

\item{cohortIds}{cohort ids.}

\item{cohortTable}{cohort Table Names}

\item{covariateSettings}{FeatureExtraction covariateSettings object}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not
truly support temp tables. To emulate temp tables,
provide a schema with write privileges where temp tables
can be created.}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}
}
\description{
Execute feature extraction
}
\details{
This function executes the cohort covariate
}
