% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatDecimalWithComma.R
\name{formatDecimalWithComma}
\alias{formatDecimalWithComma}
\title{Format a Decimal Number with Commas and Fixed Decimal Places}
\usage{
formatDecimalWithComma(number, decimalPlaces = 1, round = TRUE)
}
\arguments{
\item{number}{A numeric value to be formatted.}

\item{decimalPlaces}{An integer specifying the number of digits to display after the decimal point. Defaults to \code{1}.}

\item{round}{A logical value indicating whether the decimal portion should be rounded. If \code{FALSE},
the decimal part will be truncated instead. Defaults to \code{TRUE}.}
}
\value{
A character string representing the formatted number with commas as thousand separators for the integer part
  and a period separating the integer and decimal parts.
}
\description{
This function formats a decimal number by splitting it into an integer part and a decimal part,
formatting the integer part with commas as thousand separators, and formatting the decimal part to a fixed
number of digits. The decimal part can either be rounded or truncated based on the \code{round} parameter.
}
\details{
The function splits the number into its integer and decimal components. The integer part is formatted using
\code{formatC} with commas inserted as thousand separators. The decimal part is processed either by rounding or truncation,
then converted to a string with fixed decimal places. Finally, the two parts are concatenated with a period separator.
}
\examples{
formatDecimalWithComma(1234567.8912)
# Might return "1,234,567.9"

formatDecimalWithComma(1234567.8912, decimalPlaces = 2, round = FALSE)
# Might return "1,234,567.89"

}
