% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCountPercent.R
\name{formatIntegerWithComma}
\alias{formatIntegerWithComma}
\title{Format an Integer with Comma Separators}
\usage{
formatIntegerWithComma(number)
}
\arguments{
\item{number}{A numeric value to be formatted as an integer.}
}
\value{
A character string representing the formatted integer with commas.
}
\description{
This function formats a numeric value by first truncating any fractional part.
If the input number is a whole number (i.e., it has no fractional part), the function
inserts commas as thousand separators using \code{formatC} with \code{format = "d"} and \code{big.mark = ","}.
For non-whole numbers, only the integer portion is returned as a character string without commas.
}
\examples{
formatIntegerWithComma(1234567)
# [1] "1,234,567"

}
