% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCountPercent.R
\name{formatPercent}
\alias{formatPercent}
\title{Format a Number as a Percentage String}
\usage{
formatPercent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{A numeric value representing a proportion (e.g., 0.25 for 25%).}

\item{digits}{An integer specifying the number of digits after the decimal point. Defaults to \code{2}.}

\item{format}{A character string indicating the format to be used by \code{formatC}. Defaults to \code{"f"}.}

\item{...}{Additional arguments passed to \code{formatC} for further customization.}
}
\value{
A character string representing the formatted percentage value with a trailing percent sign.
}
\description{
This function converts a numeric value into a percentage string by multiplying the value by 100,
formatting it with a specified number of digits, and appending a percent sign ("%").
}
\details{
The function multiplies the input \code{x} by 100 to convert it to a percentage,
formats the resulting number using \code{formatC} with the specified \code{digits} and \code{format},
and finally appends a "%" sign to create a complete percentage string.
}
\examples{
formatPercent(0.1234)
# "12.34\%"

formatPercent(0.5, digits = 0)
# "50\%"

}
