% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getCohortSummary}
\alias{getCohortSummary}
\title{Check if a cohort has been instantiated.}
\usage{
getCohortSummary(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  oracleTempSchema = NULL,
  cohortTable = "cohort",
  cohortId
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortId}{The cohort definition ID used to reference the cohort in the cohort
table.}
}
\value{
Returns a data frame of cohort count, person year count, 
                             cohort_start (max and min dates), cohor_end (max and min dates)
}
\description{
Returns TRUE when a cohort has records
}
\examples{
\dontrun{
getCohortSummary(connection = connection,
                 cohortDatabaseSchema = cohortDatabaseSchema,
                 cohortId = 111)
}
}
