% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCovariateSettingsTimeWindows.R
\name{getCovariateSettingsTimeWindows}
\alias{getCovariateSettingsTimeWindows}
\title{Get Covariate Settings Time Windows}
\usage{
getCovariateSettingsTimeWindows(covariateSettings)
}
\arguments{
\item{covariateSettings}{An object (e.g., a list) containing the covariate settings. It must include the elements
\code{temporalStartDays} and \code{temporalEndDays} that define the start and end days for the time windows.}
}
\value{
A tibble with the time window definitions. The tibble contains columns \code{startDay} and \code{endDay},
  along with any additional columns from the default time windows as provided by
  \code{getFeatureExtractionDefaultTimeWindows()}.
}
\description{
This function extracts time window information from a covariate settings object. It creates a tibble
using the \code{temporalStartDays} and \code{temporalEndDays} provided in the covariate settings and performs
a left join with the default time windows from \code{getFeatureExtractionDefaultTimeWindows()}.
}
\details{
The function works by first constructing a tibble from the \code{temporalStartDays} and
\code{temporalEndDays} in the provided covariate settings. It then merges this tibble with the default
time window definitions using a left join on the \code{startDay} and \code{endDay} columns.
}
\examples{
\dontrun{
# Example covariate settings list with temporal start and end days
covariateSettings <- list(
  temporalStartDays = c(-365, -30, 1),
  temporalEndDays = c(-1, 0, 30)
)

# Retrieve the time windows based on the covariate settings
timeWindows <- getCovariateSettingsTimeWindows(covariateSettings)
print(timeWindows)
}

}
