% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRate.R
\name{getRate}
\alias{getRate}
\title{Compute rate for a cohort}
\usage{
getRate(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = NULL,
  firstOccurrenceOnly = TRUE,
  washoutPeriod = 365,
  cohortId,
  calendarPeriod = NULL,
  rateType = "incidence"
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{firstOccurrenceOnly}{Use only the first occurrence of the cohort per person?}

\item{washoutPeriod}{The minimum amount of observation time required before the occurrence
of a cohort entry. This is also used to eliminate immortal time from
the denominator.}

\item{cohortId}{The cohort definition ID used to reference the cohort in the cohort
table.}

\item{calendarPeriod}{A R-object that is an output of \code{Kala::convertDateVectorToDateSpan}.
If not provided, a default calendarPeriod will be computed based on
calendar years between cohort_start_date and cohort_end_date.}

\item{rateType}{Do you want 'incidence' or 'prevalence' for a calendarPeriod? 
Default = 'incidence'.}
}
\value{
Returns a data frame of cohort count, person year count, period begin, 
                             period end with the following stratifications: 1) no stratification, 
                             2) gender stratification, 3) age (10-year) stratification,
}
\description{
Returns rate stratified by age and gender
}
