% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRate.R
\name{getTimeSeriesMeasures}
\alias{getTimeSeriesMeasures}
\title{Returns measurement per day for cohort}
\usage{
getTimeSeriesMeasures(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cdmDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = NULL,
  washoutPeriod = 365,
  cohortId,
  asTsibble = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{washoutPeriod}{The minimum amount of observation time required before the occurrence
of a cohort entry. This is also used to eliminate immortal time from
the denominator.}

\item{cohortId}{The cohort definition ID used to reference the cohort in the cohort
table.}

\item{asTsibble}{Should the returned data frame be in tsibble format?}
}
\value{
Returns a tsibble (data frame) with keys (gender, ageGroup, 
                             washoutPeriod), 
                             index = periodBegin. The keys are parameters used in function call.
}
\description{
Returns measurement per day for cohort stratified by age and gender
}
