% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kala.R
\docType{package}
\name{Kala-package}
\alias{Kala}
\alias{Kala-package}
\title{Kala: Tools for Exploring Temporal Relationships Between Cohorts and Their Features}
\description{
The 'Kala' package provides a streamlined workflow to generate and explore temporal features for cohorts in the OMOP Common Data Model. By leveraging 'FeatureExtraction' and its 'temporalCovariateSettings', 'Kala' produces both tabular and graphical summaries of features across configurable time windows. These outputs allow easy comparisons and time-series analyses for identifying patterns, trends, or changes in clinical features over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/Kala}
  \item \url{https://github.com/OHDSI/Kala}
  \item Report bugs at \url{https://github.com/OHDSI/Kala/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{gowthamrao@gmail.com}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
