% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawTemporalCovariatesBarPlot.R
\name{rawTemporalCovariatesBarPlot}
\alias{rawTemporalCovariatesBarPlot}
\title{Create Bar Plot of temporal covariates}
\usage{
rawTemporalCovariatesBarPlot(
  covariateData,
  covariateIds = NULL,
  covariateName = "covariateName",
  covariateValueToPlot = "sumValue",
  barLabel = "report",
  barPosition = "stack",
  xAxisLabel = "Time Window",
  yAxisLabel = NULL,
  plotTitle = NULL
)
}
\arguments{
\item{covariateData}{A data frame or list containing temporal covariates. If a list, it must include a `raw` element
that is a data frame. The data frame must include the following columns:
- `covariateId` (numeric): Unique identifier for the covariate.
- `periodName` (character): Name of each time period.
- `timeId` (numeric): A numeric identifier for the time period.
- `<covariateName>` and `<covariateValueToPlot>` data, can have different names, which could be provided as function arguments.}

\item{covariateName}{The name of the covariate column in the \code{covariateData} object. Used to group covariates and draw the color legend 
Default: `"covariateName"`.}

\item{covariateValueToPlot}{The name of the value column to be drawn on the y-axis
Default: `"sumValue"`.}

\item{barLabel}{The name of the column with text information that can be depicted next to each bar.
Default: `"report"`.}

\item{barPosition}{Bar position adjustment. A string naming the position adjustment. To give the position as a string, 
strip the function name of the ggplot2::position_ prefix. For example, to use position_jitter(), give the position as "jitter".
Default: `"stack"`.}

\item{xAxisLabel}{The name of x axis. 
Default: `"Time Window"`.}

\item{yAxisLabel}{The name of y axis.
Default: `NULL` (will fallback to `covariateValueToPlot` name).}

\item{plotTitle}{The title of the plot
Default: `NULL` (no title).}

\item{covariateId}{A vector of covariate ids, that will be plotted. If `NULL`, all available covariates in the object will be plotted.
Default: `NULL`.}
}
\value{
ggplot object
}
\description{
This function draws bar plot of the raw temporal covariates returned by \code{getFeatureExtractionReportByTimeWindows} function call,
or \code{raw} part of \code{getFeatureExtractionReportByTimeWindows} with possible modifications.
It orders covariates by timeId and returns resulting bar plot object
}
\examples{
\dontrun{
# Get time windows report
timeWindowsFeatureExtractionReport <- Kala::getFeatureExtractionReportByTimeWindows(
  covariateData = covariateResultsAggregated,
  includedCovariateIds = includedCovariateIds,
  cohortId = cohortId
)
# plot temporal covariates
rawTemporalCovariatesBarPlot(timeWindowsFeatureExtractionReport)


# Filter and mutate data before plotting
covDataForPlot <- timeWindowsFeatureExtractionReport$raw |>
 dplyr::mutate(pct = scales::percent(x = averageValue, accuracy = 0.1)) |>
 dplyr::filter(timeId \%in\% c(61:85))                                            
}
# plot temporal covariates
rawTemporalCovariatesBarPlot(covDataForPlot, barLabel = "pct")

# provide different `barPosition`
rawTemporalCovariatesBarPlot(covDataForPlot, barLabel = "pct", barPosition = "dodge")


}
