% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCountPercent.R
\name{formatIntegerWithComma}
\alias{formatIntegerWithComma}
\title{Format an Integer with Comma Separators}
\usage{
formatIntegerWithComma(number)
}
\arguments{
\item{number}{A numeric value to be formatted as an integer.}
}
\value{
A character string representing the formatted integer with commas.
}
\description{
This function formats a numeric value as an integer and inserts commas as thousand separators.
It utilizes \code{formatC} with \code{format = "d"} and \code{big.mark = ","} to produce a human-readable string.
}
\examples{
formatIntegerWithComma(1234567)
# [1] "1,234,567"

}
