% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kala.R
\docType{package}
\name{Kala-package}
\alias{Kala}
\alias{Kala-package}
\title{Kala: Time series analysis for OHDSI studies}
\description{
kala takes one or more instantiated cohort as input and produces time series features, decomposition's, statistical summaries and convenient visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/Kala}
  \item \url{https://github.com/OHDSI/Kala}
  \item Report bugs at \url{https://github.com/OHDSI/Kala/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{gowthamrao@gmail.com}

Authors:
\itemize{
  \item Martijn Schuemie
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
