% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kala.R
\docType{package}
\name{Kala-package}
\alias{Kala}
\alias{Kala-package}
\title{Kala: Temporal relationships between cohort and its features}
\description{
Kala studies temporal relationship between a cohort and its features
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/Kala}
  \item \url{https://github.com/OHDSI/Kala}
  \item Report bugs at \url{https://github.com/OHDSI/Kala/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{gowthamrao@gmail.com}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
